finished separating out sub expressions
