+++ src/ast/expr/mod.rs
@@ -7,6 +7,7 @@ mod ident;
 mod r#if;
 mod literal;
 mod operator;
+mod r#return;
 mod size_of;
 mod syscall;
 mod unary;
@@ -22,6 +23,7 @@ pub use literal::*;
 pub use operator::*;
 pub use r#as::*;
 pub use r#if::*;
+pub use r#return::*;
 pub use r#while::*;
 pub use size_of::*;
 pub use syscall::*;
